requirements.txt
asgiref==3.8.1
certifi==2026.2.25
charset-normalizer==3.4.4
Django==5.2
django-dotenv==1.4.2
django-jazzmin==3.0.1
django-querycount==0.8.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.11.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.6.3
